lambda/auth/requirements.txt
pyjwt[crypto]==2.8.0
cryptography>=42.0.0
requests==2.31.0
